Add helper bridge in case of openvswitch
Since ebtables/iptables do not work on openvswitch bridges
we add a helper bridge between tap interface and ovs bridge
and connect them both with and two veth peers.

This will make MAC filtering used for private networks work
as long as nfdhcpd will be able to parse the packet mangled
and coming from the linux bridge correctly.

Lets assume the tap interface is tap1 and the ovs bridge is ovs-br.
With this patch we end up with:

tap1 -> tap1-br -> tap1-to-ovs-br -> ovs-br-to-tap3 -> ovs-br
           |                  \        /                  |
     (linux bridge)          (veth peers)           (openvswitch)

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>